First basic linker implementation
Allows users to provide their own small node modules to be included
during execution. These can be used to modify the rendering process
to achieve tailored results. Link a module using the -l parameter
(eg: -l my_module). Multiple modules can be linked, but be aware
further modules will override functionality (not extend it).

Modules must export a function to be loaded properly.
Here is a template:

    module.exports = function(marked) { /*my code*/ }

The first parameter is a reference to the marked module.